+++ src/main/scala/chiffre/inject/LfsrInjector.scala
@@ -14,8 +14,7 @@
 package chiffre.inject
 
 import chisel3._
-import chisel3.util._
-import chiffre.scan._
+import chiffre.ChiffreInjector
 
 import chiffre.{ScanField, InjectorInfo, ProbabilityBind}
