+++ src/main/scala/chiffre/inject/StuckAt.scala
@@ -14,8 +14,8 @@
 package chiffre.inject
 
 import chisel3._
-import chisel3.util._
-import chiffre.scan._
+import chisel3.util.Fill
+import chiffre.ChiffreInjector
 
 import chiffre.{ScanField, InjectorInfo}
